backend/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dotenv==1.0.0
flask==3.0.0
Flask-Cors==4.0.0
googlesearch-python==1.2.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openai==0.6.4
requests==2.31.0
soupsieve==2.5
urllib3==2.0.7
werkzeug==3.0.1
zipp==3.17.0
